Fix: simple mark issue fix
