--- frontend/src/components/meeting_arrangement/InviteUsers.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import ContinueBtn from "./buttons/ContinueBtn";
-import BackBtn from "./buttons/BackBtn";
-
-const InviteUsers: React.FC = () => {
-  return (
-    <div>
-      invite users 
-      <ContinueBtn />
-    </div>
-  )
-}
-
-export default InviteUsers
