+++ frontend/src/components/meeting_arrangement/Tide/InvitedList.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React from "react";
 import TideProfile from "./TideProfile";
 
 
@@ -16,12 +16,10 @@ const InvitedList: React.FC<{
   })
 
   return (
-    <div className="py-10 w-[40%] text-black">
-      <div className="bg-gray-200 h-[99%] rounded-3xl p-7">
-        <h2 className="font-semibold text-xl">Invited Members</h2>
-        <div className="w-[130%] flex flex-row flex-wrap">
-          {list}
-        </div>
+    <div className="bg-gray-200 rounded-3xl p-7 mt-7 text-black h-full">
+      <h2 className="font-semibold text-xl">Invited Members</h2>
+      <div className="w-[130%] flex flex-row flex-wrap">
+        {list}
       </div>
     </div>
   )
