+++ frontend/src/components/meeting_arrangement/Tide/TideSearch.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { trpc } from "../../utils/trpc";
+import { trpc } from "../../../utils/trpc";
 import TideProfile from "./TideProfile";
